# FinanceManagement/src/main/java/com/babakjan/moneybag/dto/user/AuthenticationRequest.java
package com.babakjan.moneybag.dto.user;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Size;
import lombok.*;

@Data
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class AuthenticationRequest {

    @NotNull
    @Email
    private String email;

    @NotNull
    @Size(min = 8, max = 40)
    private String password;
}