# FinanceManagement/src/main/java/com/babakjan/moneybag/dto/user/AuthenticationResponse.java
package com.babakjan.moneybag.dto.user;

import jakarta.validation.constraints.NotNull;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class AuthenticationResponse {

    private String token;

    @NotNull
    private UserDto user;
}